Fixed memory leaks on splash and main menu
Move all game state objects from stack to heap using std::unique_ptr and provide reference aliases so existing code is unchanged. Add <memory> include. Add explicit static_casts for AEFrameRateControllerGetFrameTime() and for camera Y assignment to fix type warnings/ensure consistent types. Also apply the same float cast in mainmenu.cpp and splash.cpp.

+++ Four Peaks/AlphaTemp/mainmenu.cpp
@@ -179,7 +179,7 @@ namespace game
     // -------------------------------------------------------------------------
     int MainMenu::update()
     {
-        float dt = AEFrameRateControllerGetFrameTime();
+        float dt = static_cast<float>(AEFrameRateControllerGetFrameTime());
 
         // Advance background animation
         backgroundTimer += dt;
